requirements.txt
# Backend Framework
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# AI & RAG Engine
langchain>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.1.0
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
langchain-huggingface>=0.0.3
langchain-classic>=1.0.1

# Vector Database & Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.3.1
numpy>=1.24.0

# Document Processing
pypdf>=3.17.4

# Configuration
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Utilities
openai>=1.30.0
Pillow>=10.0.0

# Testing
pytest>=7.4.0
httpx>=0.24.0
